Move list functions to own compile unit
