+++ win/libsass.vcxproj.filters
@@ -114,6 +114,9 @@
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\functions.hpp">
       <Filter>Headers</Filter>
     </ClInclude>
+    <ClInclude Include="$(LIBSASS_HEADERS_DIR)\fn_lists.hpp">
+      <Filter>Headers</Filter>
+    </ClInclude>
     <ClInclude Include="$(LIBSASS_HEADERS_DIR)\fn_colors.hpp">
       <Filter>Headers</Filter>
     </ClInclude>
@@ -285,6 +288,9 @@
       <Filter>Sources</Filter>
     </ClCompile>
     <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_utils.cpp">
+      <Filter>Sources</Filter>
+    </ClCompile>
+    <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_lists.cpp">
       <Filter>Sources</Filter>
     </ClCompile>
     <ClCompile Include="$(LIBSASS_SRC_DIR)\fn_colors.cpp">
